Expand braking fields to int32_t, refactor CAN logging, extend set_limits API
- Change braking_cmd/min/max/position/dedup fields from int16_t to int32_t for full actuator travel range
- Move extended CAN frame logging from can_twai to stepper_motor_uim2852 with richer info (CW name, node ID)
- Add stepper_uim2852_cw_name() and stepper_uim2852_rx_type() helpers to protocol header
- Track last TX payload (last_tx_data/dlc/request_ack) in motor struct under critical section
- Extend set_limits() to accept max_speed, max_accel, and enable_hw_limits flag
- Make PT frame time and FIFO low-water mark configurable via Kconfig

+++ packages/can/common/logic/control_logic/control_logic.cpp
@@ -500,7 +500,7 @@ control_step_result_t control_compute_step(node_state_t current_state, node_faul
 		// If min == max == 0, envelope is not configured; force neutral (0)
 		// instead of passing through raw planner commands.
 		int32_t clamped_steering = inputs->steering_cmd;
-		int16_t clamped_braking = inputs->braking_cmd;
+		int32_t clamped_braking = inputs->braking_cmd;
 		if (inputs->steering_min == 0 && inputs->steering_max == 0)
 		{
 			clamped_steering = 0;
@@ -515,8 +515,7 @@ control_step_result_t control_compute_step(node_state_t current_state, node_faul
 		}
 		else
 		{
-			clamped_braking =
-				(int16_t)control_clamp_command(inputs->braking_cmd, inputs->braking_min, inputs->braking_max);
+			clamped_braking = control_clamp_command(inputs->braking_cmd, inputs->braking_min, inputs->braking_max);
 		}
 
 		// Expose the clamped targets even when they have not changed so the
